requirements.txt
cupy_cuda11x==12.2.0
Cython==0.29.33
matplotlib==3.5.3
numba==0.59.1
numpy==1.21.6
pandas==1.3.5
Pillow==9.5.0
Pillow==11.3.0
trackpy==0.6.2
